requirements.txt
pillow==11.0.0
py-cord==2.6.1
PyMySQL==1.1.1
python-dotenv==1.0.1
